Add files via upload
I think this Scanner is now fully functional and updated to account for all the things Dale wanted except maybe this weird thing with how it does Unterminated Strings which may or may not be an error. I'll look in the morning. Note: ifGetNextChar should now be absolutely obsolete, but it's really useful for pinpointing bugs, so it's still in there